refactor : payment 를  facade 로 리펙토링
