+++ src/main/java/com/example/concertreservationsystem/application/usecase/PaymentUseCase.java
@@ -1,7 +1,7 @@
 package com.example.concertreservationsystem.application.usecase;
 
-import com.example.concertreservationsystem.application.user.dto.request.UserPaymentRequestDto;
-import com.example.concertreservationsystem.application.user.dto.response.UserPaymentResponseDto;
+import com.example.concertreservationsystem.application.payment.dto.request.UserPaymentRequestDto;
+import com.example.concertreservationsystem.application.payment.dto.response.UserPaymentResponseDto;
 import org.springframework.stereotype.Component;
 
 @Component
