+++ src/test/java/com/example/concertreservationsystem/SeatConcurrencyTest.java
@@ -5,7 +5,7 @@ import com.example.concertreservationsystem.domain.model.*;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
 import com.example.concertreservationsystem.domain.repo.ReservationRepository;
 import com.example.concertreservationsystem.domain.repo.UserRepository;
-import com.example.concertreservationsystem.application.user.facade.PaymentService;
+import com.example.concertreservationsystem.application.payment.facade.PaymentService;
 import com.example.concertreservationsystem.application.reservation.facade.ReservationService;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaConcertEventRepository;
 import static org.junit.jupiter.api.Assertions.*;
@@ -13,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaQueueRepository;
 import com.example.concertreservationsystem.infrastructure.persistence.JpaSeatRepository;
 import com.example.concertreservationsystem.application.reservation.dto.request.ReservationRequestDto;
-import com.example.concertreservationsystem.application.user.dto.request.UserPaymentRequestDto;
+import com.example.concertreservationsystem.application.payment.dto.request.UserPaymentRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
